[refactor] api.js 생성, ROUTES 코드 분할
